Revive examples page by fixing configuration issues
