+++ apps/examples/README.md
@@ -1,6 +1,6 @@
 # Available Examples
 
 - CSV ( complex calculation )
-- Buble Sort ( complex calculation )
+- Bubble Sort ( complex calculation )
 - External Scripts ( `removeDependencies` option example )
 - Transferable ( `transferable` option example )
