+++ apps/examples/src/pages/ExternalScripts/index.jsx
@@ -1,17 +1,30 @@
 import { WORKER_STATUS, useWorker } from '@koale/useworker'
 import React from 'react'
-import { useToasts } from 'react-toast-notifications'
+import toast from 'react-hot-toast'
 
-import bubleSort from './algorithms/bublesort'
+import sortDates from './algorithms/sortDates'
 
-const numbers = [...Array(50000)].map(() => Math.floor(Math.random() * 1000000))
+const dates = [...Array(100000)].map(
+  () => new Date(1995, Math.floor(Math.random() * 2000), 6, 2),
+)
 
 function App() {
-  const { addToast } = useToasts()
-
   const [sortStatus, setSortStatus] = React.useState(false)
+
   const [sortWorker, { status: sortWorkerStatus, kill: killWorker }] =
-    useWorker(bubleSort)
+    useWorker(sortDates, {
+      autoTerminate: false, // you should manually kill the worker using "killWorker()"
+      remoteDependencies: [
+        'https://cdnjs.cloudflare.com/ajax/libs/date-fns/1.30.1/date_fns.js',
+      ],
+    })
+
+  React.useEffect(
+    () => () => {
+      killWorker() // [UN-MOUNT] Since autoTerminate: false we need to kill the worker manually (recommended)
+    },
+    [killWorker],
+  )
 
   React.useEffect(() => {
     console.log('WORKER:', sortWorkerStatus)
@@ -19,16 +32,16 @@ function App() {
 
   const onSortClick = () => {
     setSortStatus(true)
-    const result = bubleSort(numbers)
+    const result = sortDates(dates)
     setSortStatus(false)
-    addToast('Finished: Sort', { appearance: 'success' })
-    console.log('Buble Sort', result)
+    toast.success('Finished: Sort')
+    console.log('Bubble Sort', result)
   }
 
   const onWorkerSortClick = () => {
-    sortWorker(numbers).then((result) => {
-      console.log('Buble Sort useWorker()', result)
-      addToast('Finished: Sort using useWorker.', { appearance: 'success' })
+    sortWorker(dates).then((result) => {
+      console.log('Bubble Sort useWorker()', result)
+      toast.success('Finished: Sort using useWorker.')
     })
   }
 
@@ -41,7 +54,7 @@ function App() {
           className="App-button"
           onClick={() => onSortClick()}
         >
-          {sortStatus ? `Loading...` : `Buble Sort`}
+          {sortStatus ? `Loading...` : `Sort Dates`}
         </button>
         <button
           type="button"
@@ -51,7 +64,7 @@ function App() {
         >
           {sortWorkerStatus === WORKER_STATUS.RUNNING
             ? `Loading...`
-            : `Buble Sort useWorker()`}
+            : `Sort Dates useWorker()`}
         </button>
         {sortWorkerStatus === WORKER_STATUS.RUNNING ? (
           <button
