+++ apps/examples/vite.config.js
@@ -3,7 +3,7 @@ import { defineConfig } from 'vite'
 
 export default defineConfig({
   plugins: [react()],
-  root: 'src',
+  root: '.',
   build: {
     outDir: '../dist',
   },
